requirements.txt
flask==3.0.0
flask-cors==4.0.0
pillow==10.0.0
requests==2.31.0
numpy==1.24.3
gunicorn==21.2.0
